react: extend React.Component, closes #13
This makes the React adapter identical to the Preact one, since Preact
is API compatible with React core.
