+++ react.js
@@ -6,31 +6,38 @@ function toReact (Component, react) {
   assert.equal(typeof Component, 'function', 'nanocomponent-adapters/react: component should be type function')
   assert.equal(typeof react, 'object', 'nanocomponent-adapters/react: react should be type object')
 
-  var clx = react.createClass({
-    getInitialState: function () {
-      this.node = null
-      this.comp = new Component()
-      return {}
-    },
-    shouldComponentUpdate: function (nextProps) {
-      return false
-    },
-    componentWillReceiveProps: function (nwProps) {
-      if (this.comp.element) this.comp.render(nwProps)
-    },
-    componentDidMount: function () {
-      if (!this.comp.element) {
-        var el = this.comp.render(this.props)
-        this.node.appendChild(el)
-      }
-    },
-    setRef: function (_node) {
-      this.node = _node
-    },
-    render: function () {
-      return react.createElement('div', { ref: this.setRef })
-    }
-  })
+  function Clx () {
+    react.Component.apply(this, arguments)
+    this.comp = new Component()
+    this.node = null
+    this.setRef = this.setRef.bind(this)
+  }
 
-  return clx
+  Clx.prototype = Object.create(react.Component.prototype)
+  Clx.prototype.constructor = react.Component
+
+  Clx.prototype.componentDidMount = function componentDidMount () {
+    if (!this.comp.element) {
+      var el = this.comp.render(this.props)
+      this.node.appendChild(el)
+    }
+  }
+
+  Clx.prototype.setRef = function (_node) {
+    this.node = _node
+  }
+
+  Clx.prototype.componentWillReceiveProps = function componentWillReceiveProps (props) {
+    if (this.comp.element) this.comp.render(props)
+  }
+
+  Clx.prototype.shouldComponentUpdate = function shouldComponentUpdate () {
+    return false
+  }
+
+  Clx.prototype.render = function render (props) {
+    return react.createElement('div', { ref: this.setRef })
+  }
+
+  return Clx
 }
